move exporter package to pkg/exporter
